update newsletter error
