+++ src/index.html
@@ -6,7 +6,6 @@
     <title>Sleep Outside | Home</title>
     <link rel="stylesheet" href="/css/style.css" />
     <script src="js/main.js" type="module"></script>
-    <script src="/js/newsletter.js" defer></script>
   </head>
   <body>
     <header id="main-header" class="divider"></header>
