+++ src/js/main.js
@@ -1,4 +1,6 @@
 // import { displayCounter } from "./cartCount.mjs";
-import { loadHeaderFooter } from "./utils.mjs";
+import { loadHeaderFooter, loadNewsletter } from "./utils.mjs";
 
 loadHeaderFooter();
+
+loadNewsletter();
