+++ src/js/newsletter.js
@@ -1,28 +1,33 @@
-const newsletterButton = document.getElementById("newsletter-button");
-const newsletterModal = document.getElementById("newsletter-modal");
-const closeButton = document.getElementsByClassName("close")[0];
-const form = document.getElementById("newsletter-form");
-const thankYou = document.getElementById("thank-you");
+// moved to utils
 
-newsletterButton.addEventListener("click", function () {
-  newsletterModal.style.display = "block";
-});
+// let loadNewsletter = document.addEventListener("DOMContentLoaded", function() {
 
-closeButton.addEventListener("click", function () {
-  newsletterModal.style.display = "none";
-});
-
-window.addEventListener("click", function (event) {
-  if (event.target == newsletterModal) {
-    newsletterModal.style.display = "none";
-  }
-});
-
-form.addEventListener("submit", function (event) {
-  event.preventDefault();
-  const fname = document.getElementById("news-fname").value;
-  document.getElementById(
-    "thank-you"
-  ).innerHTML = `Thank you for subscribing, ${fname}!`;
-  thankYou.style.display = "block";
-});
+//   const newsletterButton = document.getElementById("newsletter-button");
+//   const newsletterModal = document.getElementById("newsletter-modal");
+//   const closeButton = document.getElementsByClassName("close")[0];
+//   const form = document.getElementById("newsletter-form");
+//   const thankYou = document.getElementById("thank-you");
+  
+//   newsletterButton.addEventListener("click", function () {
+//     newsletterModal.style.display = "block";
+//   });
+  
+//   closeButton.addEventListener("click", function () {
+//     newsletterModal.style.display = "none";
+//   });
+  
+//   window.addEventListener("click", function (event) {
+//     if (event.target == newsletterModal) {
+//       newsletterModal.style.display = "none";
+//     }
+//   });
+  
+//   form.addEventListener("submit", function (event) {
+//     event.preventDefault();
+//     const fname = document.getElementById("news-fname").value;
+//     document.getElementById(
+//       "thank-you"
+//     ).innerHTML = `Thank you for subscribing, ${fname}!`;
+//     thankYou.style.display = "block";
+//   });
+// });
